+++ app/src/main/java/de/readeckapp/ui/detail/BookmarkDetailViewModel.kt
@@ -6,14 +6,18 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import de.readeckapp.domain.BookmarkRepository
+import de.readeckapp.domain.model.Template
+import de.readeckapp.domain.model.Theme
 import de.readeckapp.domain.usecase.UpdateBookmarkUseCase
 import de.readeckapp.io.AssetLoader
+import de.readeckapp.io.prefs.SettingsDataStore
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
@@ -24,7 +28,6 @@ import timber.log.Timber
 import java.text.DateFormat
 import java.util.Date
 import javax.inject.Inject
-import kotlin.io.encoding.Base64
 import kotlin.io.encoding.ExperimentalEncodingApi
 
 @HiltViewModel
@@ -32,6 +35,7 @@ class BookmarkDetailViewModel @Inject constructor(
     private val updateBookmarkUseCase: UpdateBookmarkUseCase,
     private val bookmarkRepository: BookmarkRepository,
     private val assetLoader: AssetLoader,
+    private val settingsDataStore: SettingsDataStore,
     savedStateHandle: SavedStateHandle
 ) : ViewModel() {
     private val _navigationEvent = MutableStateFlow<NavigationEvent?>(null)
@@ -44,56 +48,47 @@ class BookmarkDetailViewModel @Inject constructor(
     val shareIntent: StateFlow<Intent?> = _shareIntent.asStateFlow()
 
     private val bookmarkId: String? = savedStateHandle["bookmarkId"]
-    private val htmlTemplate = flow {
-        emit(assetLoader.loadAsset("html_template.html"))
+    private val template: Flow<Template?> = settingsDataStore.themeFlow.map {
+        it?.let {
+            Theme.valueOf(it)
+        } ?: Theme.SYSTEM
+    }.map {
+        when (it) {
+            Theme.DARK -> assetLoader.loadAsset(Template.DARK_TEMPLATE_FILE)?.let { Template.SimpleTemplate(it) }
+            Theme.LIGHT -> assetLoader.loadAsset(Template.LIGHT_TEMPLATE_FILE)?.let { Template.SimpleTemplate(it) }
+            Theme.SYSTEM -> {
+                val light = assetLoader.loadAsset(Template.LIGHT_TEMPLATE_FILE)
+                val dark = assetLoader.loadAsset(Template.DARK_TEMPLATE_FILE)
+                if (!light.isNullOrBlank() && !dark.isNullOrBlank()) {
+                    Template.DynamicTemplate(light = light, dark = dark)
+                } else null
+            }
+        }
     }
     private val updateState = MutableStateFlow<UpdateBookmarkState?>(null)
 
     @OptIn(ExperimentalEncodingApi::class)
     val uiState = combine(
         bookmarkRepository.observeBookmark(bookmarkId!!),
-        htmlTemplate,
-        updateState
-    ) { bookmark, htmlTemplate, updateState ->
+        updateState,
+        template
+    ) { bookmark, updateState, template ->
         if (bookmark == null) {
             Timber.e("Error loading bookmark [bookmarkId=$bookmarkId]")
             UiState.Error
-        } else if (htmlTemplate == null) {
-            Timber.e("Error loading htmlTemplate")
+        } else if (template == null) {
+            Timber.e("Error loading template(s)")
             UiState.Error
         } else {
-            val content = when (bookmark.type) {
-                de.readeckapp.domain.model.Bookmark.Type.Picture -> {
-                    htmlTemplate.replace("%s", """<img src="${bookmark.image.src}"/>""")
-                }
-
-                de.readeckapp.domain.model.Bookmark.Type.Video -> {
-                    bookmark.articleContent?.let {
-                        htmlTemplate.replace("%s", it)
-                    }
-                }
-
-                de.readeckapp.domain.model.Bookmark.Type.Article -> {
-                    bookmark.articleContent?.let {
-                        htmlTemplate.replace("%s", it)
-                    }
-                }
-            }
-            val encodedHtml = content?.let {
-                Base64.withPadding(Base64.PaddingOption.ABSENT).encode(it.toByteArray())
-            }
-
             UiState.Success(
                 bookmark = Bookmark(
                     url = bookmark.url,
                     title = bookmark.title,
-                    encodedHtmlContent = encodedHtml,
                     authors = bookmark.authors,
                     createdDate = formatLocalDateTimeWithDateFormat(bookmark.created),
                     bookmarkId = bookmarkId,
                     siteName = bookmark.siteName,
                     imgSrc = bookmark.image.src,
-                    htmlContent = content,
                     isFavorite = bookmark.isMarked,
                     isArchived = bookmark.isArchived,
                     isRead = bookmark.isRead(),
@@ -101,9 +96,11 @@ class BookmarkDetailViewModel @Inject constructor(
                         is de.readeckapp.domain.model.Bookmark.Type.Article -> Bookmark.Type.ARTICLE
                         is de.readeckapp.domain.model.Bookmark.Type.Picture -> Bookmark.Type.PHOTO
                         is de.readeckapp.domain.model.Bookmark.Type.Video -> Bookmark.Type.VIDEO
-                    }
+                    },
+                    articleContent = bookmark.articleContent
                 ),
-                updateBookmarkState = updateState
+                updateBookmarkState = updateState,
+                template = template
             )
         }
     }
@@ -214,7 +211,7 @@ class BookmarkDetailViewModel @Inject constructor(
     }
 
     sealed class UiState {
-        data class Success(val bookmark: Bookmark, val updateBookmarkState: UpdateBookmarkState?) :
+        data class Success(val bookmark: Bookmark, val updateBookmarkState: UpdateBookmarkState?, val template: Template) :
             UiState()
 
         data object Loading : UiState()
@@ -224,20 +221,48 @@ class BookmarkDetailViewModel @Inject constructor(
     data class Bookmark(
         val url: String,
         val title: String,
-        val encodedHtmlContent: String?,
         val authors: List<String>,
         val createdDate: String,
         val bookmarkId: String,
         val siteName: String,
         val imgSrc: String,
-        val htmlContent: String?,
         val isFavorite: Boolean,
         val isArchived: Boolean,
         val isRead: Boolean,
-        val type: Type
+        val type: Type,
+        val articleContent: String?
     ) {
         enum class Type {
             ARTICLE, PHOTO, VIDEO
+        }
+        fun getContent(template: Template, isDark: Boolean): String? {
+            val htmlTemplate = when (template) {
+                is Template.SimpleTemplate -> template.template
+                is Template.DynamicTemplate -> {
+                    if (isDark) {
+                        template.dark
+                    } else {
+                        template.light
+                    }
+                }
+            }
+            return when (type) {
+                Type.PHOTO -> {
+                    htmlTemplate.replace("%s", """<img src="$imgSrc"/>""")
+                }
+
+                Type.VIDEO -> {
+                    articleContent?.let {
+                        htmlTemplate.replace("%s", it)
+                    }
+                }
+
+                Type.ARTICLE -> {
+                    articleContent?.let {
+                        htmlTemplate.replace("%s", it)
+                    }
+                }
+            }
         }
     }